Move exportTransformers helper method to cmd root
